backend/requirements.txt
django>=4.2
djangorestframework>=3.14
corsheaders>=4.0
django-filter>=23.0
Pillow>=10.0
psycopg2-binary>=2.9
python-decouple>=3.8
